Relocate more TreeViewReduxTypes
