+++ src/components/App/DefaultState.ts
@@ -4,7 +4,7 @@ import { defaultState as currentProjectState } from "components/Project/ProjectR
 import { defaultState as exportProjectState } from "components/ProjectExport/Redux/ExportProjectReduxTypes";
 import { defaultState as createProjectState } from "components/ProjectScreen/CreateProject/Redux/CreateProjectReduxTypes";
 import { defaultState as pronunciationsState } from "components/Pronunciations/Redux/PronunciationsReduxTypes";
-import { defaultState as treeViewState } from "components/TreeView/Redux/TreeViewReducer";
+import { defaultState as treeViewState } from "components/TreeView/Redux/TreeViewReduxTypes";
 import { defaultState as characterInventoryState } from "goals/CharInventoryCreation/Redux/CharacterInventoryReducer";
 import { defaultState as mergeDuplicateGoal } from "goals/MergeDupGoal/Redux/MergeDupReducer";
 import { defaultState as reviewEntriesState } from "goals/ReviewEntries/ReviewEntriesComponent/Redux/ReviewEntriesReduxTypes";
