+++ src/components/TreeView/Redux/TreeViewReduxTypes.tsx
@@ -15,4 +15,16 @@ export interface TreeViewAction {
   language?: string;
 }
 
+export interface TreeViewState {
+  currentDomain: SemanticDomainTreeNode;
+  language: string;
+  open: boolean;
+}
+
 export const defaultTreeNode = newSemanticDomainTreeNode("Sem");
+
+export const defaultState: TreeViewState = {
+  language: "",
+  open: false,
+  currentDomain: defaultTreeNode,
+};
